Move handle_java_exception up to Py4JBackend
