+++ hail/python/hail/backend/py4j_backend.py
@@ -1,8 +1,40 @@
 import abc
 
+import py4j
+
+import hail
 from hail.ir.renderer import CSERenderer
-from hail.utils.java import Env
+from hail.utils.java import FatalError, Env
 from .backend import Backend
+
+
+def handle_java_exception(f):
+    def deco(*args, **kwargs):
+        import pyspark
+        try:
+            return f(*args, **kwargs)
+        except py4j.protocol.Py4JJavaError as e:
+            s = e.java_exception.toString()
+
+            # py4j catches NoSuchElementExceptions to stop array iteration
+            if s.startswith('java.util.NoSuchElementException'):
+                raise
+
+            tpl = Env.jutils().handleForPython(e.java_exception)
+            deepest, full, error_id = tpl._1(), tpl._2(), tpl._3()
+
+            if error_id != -1:
+                raise FatalError('Error summary: %s' % (deepest,), error_id) from None
+            else:
+                raise FatalError('%s\n\nJava stack trace:\n%s\n'
+                                 'Hail version: %s\n'
+                                 'Error summary: %s' % (deepest, full, hail.__version__, deepest), error_id) from None
+        except pyspark.sql.utils.CapturedException as e:
+            raise FatalError('%s\n\nJava stack trace:\n%s\n'
+                             'Hail version: %s\n'
+                             'Error summary: %s' % (e.desc, e.stackTrace, hail.__version__, e.desc)) from None
+
+    return deco
 
 
 class Py4JBackend(Backend):
