+++ hail/python/hail/backend/spark_backend.py
@@ -19,43 +19,13 @@ from hail.ir import JavaIR
 from hail.table import Table
 from hail.matrixtable import MatrixTable
 
-from .py4j_backend import Py4JBackend
+from .py4j_backend import Py4JBackend, handle_java_exception
 from ..hail_logging import Logger
 
 
 if pyspark.__version__ < '3' and sys.version_info > (3, 8):
     raise EnvironmentError('Hail with spark {} requires Python 3.6 or 3.7, found {}.{}'.format(
         pyspark.__version__, sys.version_info.major, sys.version_info.minor))
-
-
-def handle_java_exception(f):
-    def deco(*args, **kwargs):
-        import pyspark
-        try:
-            return f(*args, **kwargs)
-        except py4j.protocol.Py4JJavaError as e:
-            s = e.java_exception.toString()
-
-            # py4j catches NoSuchElementExceptions to stop array iteration
-            if s.startswith('java.util.NoSuchElementException'):
-                raise
-
-            tpl = Env.jutils().handleForPython(e.java_exception)
-            deepest, full, error_id = tpl._1(), tpl._2(), tpl._3()
-
-            if error_id != -1:
-                raise FatalError('Error summary: %s' % (deepest,), error_id) from None
-            else:
-                raise FatalError('%s\n\nJava stack trace:\n%s\n'
-                                 'Hail version: %s\n'
-                                 'Error summary: %s' % (deepest, full, hail.__version__, deepest), error_id) from None
-        except pyspark.sql.utils.CapturedException as e:
-            raise FatalError('%s\n\nJava stack trace:\n%s\n'
-                             'Hail version: %s\n'
-                             'Error summary: %s' % (e.desc, e.stackTrace, hail.__version__, e.desc)) from None
-
-    return deco
-
 
 _installed = False
 _original = None
